refactor(cli): split subcommand handlers out of app.go (#120)
