+++ internal/cli/status.go
@@ -1,0 +1,125 @@
+package cli
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os"
+	"path/filepath"
+	"sort"
+	"strings"
+
+	"dir2mcp/internal/model"
+)
+
+func (a *App) runStatus(ctx context.Context, global globalOptions, args []string) int {
+	if len(args) > 0 {
+		writef(a.stderr, "status command does not accept arguments: %s\n", strings.Join(args, " "))
+		return exitConfigInvalid
+	}
+
+	cfg, err := loadConfigWithGlobalOptions(global)
+	if err != nil {
+		writef(a.stderr, "load config: %v\n", err)
+		return exitConfigInvalid
+	}
+	if strings.TrimSpace(cfg.StateDir) == "" {
+		cfg.StateDir = filepath.Join(".", ".dir2mcp")
+	}
+
+	snapshotPath := filepath.Join(cfg.StateDir, "corpus.json")
+	snapshot, err := readCorpusSnapshot(snapshotPath)
+	source := "corpus_json"
+	if err != nil {
+		metaPath := filepath.Join(cfg.StateDir, "meta.sqlite")
+		if _, statErr := os.Stat(metaPath); statErr != nil {
+			if errors.Is(statErr, os.ErrNotExist) {
+				writeln(a.stderr, "no state found in .dir2mcp; run: dir2mcp up")
+				return exitGeneric
+			}
+			writef(a.stderr, "read state: %v\n", statErr)
+			return exitGeneric
+		}
+
+		st := a.storeForConfig(cfg)
+		defer func() { _ = st.Close() }()
+		if initErr := st.Init(ctx); initErr != nil && !errors.Is(initErr, model.ErrNotImplemented) {
+			writef(a.stderr, "initialize metadata store: %v\n", initErr)
+			return exitIndexLoadFailure
+		}
+		emitter := newNDJSONEmitter(a.stdout, global.jsonOutput)
+		snapshot, err = buildCorpusSnapshot(ctx, st, nil, a.stderr, emitter)
+		if err != nil {
+			writef(a.stderr, "build status snapshot: %v\n", err)
+			return exitGeneric
+		}
+		source = "computed"
+	}
+
+	if global.jsonOutput {
+		payload := map[string]interface{}{
+			"source":    source,
+			"state_dir": cfg.StateDir,
+			"snapshot":  snapshot,
+		}
+		if err := emitJSON(a.stdout, payload); err != nil {
+			writef(a.stderr, "encode status json: %v\n", err)
+			return exitGeneric
+		}
+		return exitSuccess
+	}
+
+	if global.quiet {
+		return exitSuccess
+	}
+	s := a.sty(false)
+	writeln(a.stdout)
+	writeln(a.stdout, s.kv("State", cfg.StateDir))
+	writeln(a.stdout, s.kv("Source", source))
+	writeln(a.stdout, s.kv("Timestamp", snapshot.Timestamp))
+	writeln(a.stdout)
+
+	runningLabel := s.dim("stopped")
+	if snapshot.Indexing.Running {
+		runningLabel = s.Green.Render("running")
+	}
+
+	writef(a.stdout, "  %s  %s  %s\n", s.sectionHeader("Indexing"), s.dim("mode="+snapshot.Indexing.Mode), runningLabel)
+	writef(a.stdout, "    %s  %s  %s  %s\n",
+		s.stat("scanned", snapshot.Indexing.Scanned),
+		s.stat("indexed", snapshot.Indexing.Indexed),
+		s.stat("skipped", snapshot.Indexing.Skipped),
+		s.stat("deleted", snapshot.Indexing.Deleted),
+	)
+	writef(a.stdout, "    %s  %s  %s  %s",
+		s.stat("reps", snapshot.Indexing.Representations),
+		s.stat("chunks", snapshot.Indexing.ChunksTotal),
+		s.stat("embedded", snapshot.Indexing.EmbeddedOK),
+		s.stat("unknown", snapshot.Indexing.Unknown),
+	)
+	if snapshot.Indexing.Errors > 0 {
+		writef(a.stdout, "  %s", s.Red.Render(fmt.Sprintf("errors=%d", snapshot.Indexing.Errors)))
+	} else {
+		writef(a.stdout, "  %s", s.stat("errors", snapshot.Indexing.Errors))
+	}
+	writeln(a.stdout)
+	writeln(a.stdout)
+
+	writef(a.stdout, "  %s  %s  %s\n",
+		s.sectionHeader("Documents"),
+		s.stat("total", snapshot.TotalDocs),
+		s.stat("code_ratio", fmt.Sprintf("%.4f", snapshot.CodeRatio)),
+	)
+	if len(snapshot.DocCounts) > 0 {
+		keys := make([]string, 0, len(snapshot.DocCounts))
+		for key := range snapshot.DocCounts {
+			keys = append(keys, key)
+		}
+		sort.Strings(keys)
+		for _, key := range keys {
+			writef(a.stdout, "    %s\n", s.stat(key, snapshot.DocCounts[key]))
+		}
+	}
+	writeln(a.stdout)
+	return exitSuccess
+}
